Make compilation tests local-only to prevent CI OOM errors
Added --local-tests flag to skip memory-intensive compilation tests in CI
while allowing them to run locally for development. This prevents error
code 137 (OOM kill) in GitHub Actions.

Changes:
- Added --local-tests pytest option
- Created @pytest.mark.local_only marker for memory-intensive tests
- Applied marker to all compilation tests
- Cache clearing fixture remains in test_compilation.py only
- Updated tests/README.md with documentation

Usage:
- CI runs: compilation tests are skipped by default
- Local runs: use --local-tests flag to include compilation tests
